requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
temporalio==1.8.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
prometheus-client==0.20.0
pyyaml==6.0.2
sympy==1.13.2
pytest==8.3.3
